+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Changelog
 Next
 ----
 
+2018.11.07.0
+------------
+
 - Add ``dcos-docker create-loopback-sidecar`` and ``dcos-docker destroy-loopback-sidecar`` commands to provide unformatted block devices to DC/OS.
 - Add ``dcos-docker clean`` command to clean left over artifacts.
 - Backwards incompatible change: Changed names of VPN containers on macOS.